pyproject.toml
[project]
name = "api-alert-system"
version = "0.1.0"
description = "Stock Alert System with Telegram Notifications and MCP Integration"
requires-python = ">=3.8"
dependencies = [
    "yfinance>=0.2.28",
    "python-telegram-bot>=20.0",
    "schedule>=1.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "fastmcp>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
alert-bot = "api_alert_system.core.alert_bot:main"
setup-db = "scripts.setup_db:main"
inspect-db = "scripts.inspect_db:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api_alert_system"] 